Add warning levels; Add spdlog as sources instead of submodule

+++ OctreeApi/OctreeLibrary/include/time_tracker.cuh
@@ -45,7 +45,7 @@ public:
         kernelTimers.emplace_back (timer, measurement);
         if (kernelOrder.find (measurement) == kernelOrder.end ())
         {
-            kernelOrder[measurement] = kernelTimings.size ();
+            kernelOrder[measurement] = static_cast<int>(kernelTimings.size ());
             kernelTimings.emplace_back (measurement, TimingProps{0.f, 0});
         }
     }
